requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
anyio==3.6.2
asgiref==3.6.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1.2
coverage==7.2.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.68.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.1.2
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.1.2
orjson==3.8.8
packaging==23.0
pluggy==1.0.0
promise==2.3
py==1.11.0
pydantic==1.10.6
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
Rx==1.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.14.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
ujson==4.3.0
urllib3==1.26.15
uvicorn==0.13.4
uvloop==0.17.0
watchgod==0.8.2
websockets==8.1
